Updated images + connector version Started working on script
