+++ src/run_download_path.py
@@ -1,0 +1,29 @@
+# Copyright 2023 Universität Tübingen, DKFZ and EMBL
+# for the German Human Genome-Phenome Archive (GHGA)
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import asyncio
+
+import crypt4gh.header
+import crypt4gh.keys
+import crypt4gh.lib
+from ghga_connector.cli import upload
+from ghga_event_schemas import pydantic_ as event_schemas
+
+
+from src.commons import BASE_DIR, CONFIG, FILE_SIZE
+
+
+def download_file(file_id: str):
+    """Download a file"""
